Disable defrag test for pubsub
When iothreads are enabled, the memory allocation for channel names subscribed to before this point was previously handled in the io threads. Now, it is always allocated in the main thread. However, key-value pairs are still allocated in io threads. As a result, tests no longer cause fragmentation between the two.